Tests/LibWeb/Text/expected/Compression/CompressionStream.txt

format=deflate: eJwLT83JUchIzcnJV0grykzNSylWBABGEQb1
format=deflate-raw: C0/NyVHISM3JyVdIK8pMzUspVgQA
format=gzip: H4sIAAAAAAADAwtPzclRyEjNyclXSCvKTM1LKVYEAHN0w4sTAAAA
